adds npm_package/lib to node_path
I had to add this to the NODE_PATH to finish the installation. Before I did, I would get an error telling me my NODE_PATH and npm root were different and to make this change to fix that. It was the global installation of yo that gave this error message.